refactor markdown component
